requirements.txt
# Updated requirements.txt with compatible versions
setuptools>=68.0.0
wheel>=0.41.0
numpy>=1.26.0
streamlit==1.29.0
sentence-transformers==2.2.2
faiss-cpu>=1.10.0
rank-bm25==0.2.2
google-generativeai==0.3.2
pandas>=2.0.3
python-dotenv>=1.0.0
huggingface-hub==0.20.3  # Add this line
transformers==4.36.2      # Add this line
